fix 32-bit

+++ src/context.jl
@@ -176,7 +176,7 @@ end
     debug && println("header is: $header, datarow computed as: $datarow")
     # getsource will turn any input into a `AbstractVector{UInt8}`
     buf, pos, len = getsource(source)
-    if len > 2^42
+    if len > Int64(2)^42
         throw(ArgumentError("delimited source to parse too large; must be < $(2^42) bytes"))
     end
     # skip over initial BOM character, if present
